# src/main/resources/CalculatorSCF-0.0.1-swagger.yaml
swagger: "2.0"
info:
  description: SCF Calculator Api
  version: "0.0.1"
  title: SCF Calculator Api
  termsOfService: urn:tos
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  contact:
      name: Team
      email: team@gruposantander.com
  x-santander-catalogation: 
      bian-landscape-version: "8"
      bian-business-area: “Business Support”
      bian-business-domain: “Document Management and Archive”
      bian-service-domain: "Document Services" 
host: "localhost:8090"
basePath: "/CalculatorSCF"
schemes:
  - "http"
  - "https"
securityDefinitions:
  JWTProfile:
    type: "oauth2"
    description: "OAuth Grant type, JWT Profile"
    flow: "implicit"
    scopes:
     sce: "sce"
    authorizationUrl: "https://tbd.santander.es"
    x-tokenIntrospect:
      url: "$(sca-security-url)"
  X-Santander-Client-Id:
    type: "apiKey"
    in: "header"
    name: "X-Santander-Client-Id"
    description: "Client id header."
tags:
  - name: "gqs-calculator"
    description: "Calculator API"
paths:
  /gqsApplication/calculator/{country}/v1/products:
    post:
      tags:
        - gqs-getProducts
      summary: Get products
      description: GQS Products information
      operationId: requestProducts
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/AcceptParam'
        - $ref: '#/parameters/ContentTypeParam'
        - $ref: '#/parameters/ParentSpanIdParam'
        - $ref: '#/parameters/SampledParam'
        - $ref: '#/parameters/SpanIdParam'
        - $ref: '#/parameters/TraceIdParam'
        - $ref: '#/parameters/SantanderClientIdParam'
        - $ref: '#/parameters/CountryParam'  
        - in: body
          name: InputParams
          description: input Data
          required: true
          schema:
            $ref: '#/definitions/InputParamsSchema'
        - name: BusinessUnit
          in: query
          description: BusinessUnit
          required: true
          type: string
          allowEmptyValue: false
        - name: Version
          in: query
          description: Version
          required: true
          type: string
          allowEmptyValue: false
      responses:
        '200':
          description: 200 OK
          schema:
            $ref: '#/definitions/ProductsResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ResponseError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Error Server
      security:
        - JWTProfile:
            - sce
      deprecated: false
  /gqsApplication/calculator/{country}/v1/calculate:
    post:
      tags:
        - gqs-calculate
      summary: Calculate one quote
      description: Get example loan cuotes for dealer
      operationId: getSimulation
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/AcceptParam'
        - $ref: '#/parameters/ContentTypeParam'
        - $ref: '#/parameters/ParentSpanIdParam'
        - $ref: '#/parameters/SampledParam'
        - $ref: '#/parameters/SpanIdParam'
        - $ref: '#/parameters/TraceIdParam'
        - $ref: '#/parameters/SantanderClientIdParam'
        - $ref: '#/parameters/PasswordParam'
        - $ref: '#/parameters/UserParam'
        - $ref: '#/parameters/CountryParam' 
        - in: body
          name: GQS
          description: GQS data
          required: true
          schema:
            $ref: '#/definitions/GQSSchema'
      responses:
        '200':
          description: 200 OK
          schema:
            $ref: '#/definitions/QuotingScheme'
        '201':
          description: Created
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ResponseError'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Error Server
      security:
        - JWTProfile:
            - sce
      deprecated: false
      
definitions:
  InputParamsSchema:
    type: object
    properties:
      DealerCode:
        type: string
      User:
        type: string
      Password:
        type: string
      CapCode:
        type: string
      VIN:
        type: string
      LicensePlate:
        type: string
  GQSSchema:
    type: object
    properties:
      Dealer:
        type: string
      Product:
        type: string
      RetailPrice:
        type: string
      DownPayment:
        type: string
      OtherLoansAmount:
        type: string
      Balloon:
        type: string
      Period:
        type: string
      Instalment:
        type: string
      InterestRate:
        type: string
      DeferredPeriod:
        type: string
      Mileage:
        type: string
      InsurancePrime:
        type: string
      FirstInstalmentDate:
        type: string
      InterestDate:
        type: string
      FeeType:
        type: string
      AssetsCode:
        type: string
        $ref: '#/definitions/AssetsCode'
      Insurances:
        $ref: '#/definitions/Insurances'
  CapCodes:
    type: object
    properties:
      CapCode:
        type: array
        items:
          type: string
    title: CapCodes
  Error:
    type: object
    title: Error
    additionalProperties:
      type: object
  Fees:
    type: object
    properties:
      fee_type_id:
        type: string
      fixed_fee:
        type: string
      max_amount:
        type: string
      min_amount:
        type: string
      payment_id:
        type: string
      pct_max_fee:
        type: string
    title: Fees
  FinanceProduct:
    type: object
    properties:
      Product:
        type: array
        items:
          $ref: '#/definitions/Product'
    title: FinanceProduct
  Product:
    type: object
    properties:
      CapCodes:
        $ref: '#/definitions/CapCodes'
      DefaultCurrency:
        type: string
      DefaultRate:
        type: string
      ManufacturerDepositContribution:
        type: string
      MarketIso:
        type: string
      MaxPayment:
        type: string
      ProductDesc:
        type: string
      ProductEndDate:
        type: string
      ProductFees:
        $ref: '#/definitions/ProductFees'
      ProductId:
        type: string
      ProductOnline:
        type: string
      ProductStartDate:
        type: string
      ProductTypeDesc:
        type: string
      ProductTypeId:
        type: string
      ProductValidations:
        $ref: '#/definitions/ProductValidations'
    title: Product
  ProductFees:
    type: object
    properties:
      Fees:
        $ref: '#/definitions/Fees'
    title: ProductFees
  ProductValidations:
    type: object
    properties:
      Validation:
        $ref: '#/definitions/Validation'
    title: ProductValidations
  ProductsResponse:
    type: object
    properties:
      FinanceProduct:
        $ref: '#/definitions/FinanceProduct'
      XmlData:
        type: string
    title: ProductsResponse
  QuoteTable:
    type: object
    properties:
      APR:
        type: string
      Capital:
        type: string
      IRR:
        type: string
      IRR2:
        type: string
      InsuranceAmount:
        type: string
      MonthInterest:
        type: string
      MonthQuote:
        type: string
      QuoteID:
        type: string
    title: QuoteTable
  QuotingScheme:
    type: object
    properties:
      QuoteTable:
        type: array
        items:
          $ref: '#/definitions/QuoteTable'
      SummaryTable:
        type: array
        items:
          $ref: '#/definitions/SummaryTable'
      TotalTable:
        $ref: '#/definitions/TotalTable'
      XmlData:
        type: string
      text_format:
        type: array
        items:
          $ref: '#/definitions/TextFormat'
    title: QuotingScheme
  ResponseError:
    type: object
    properties:
      debugMessage:
        type: string
      message:
        type: string
      status:
        type: string
        enum:
          - 100 CONTINUE
          - 101 SWITCHING_PROTOCOLS
          - 102 PROCESSING
          - 103 CHECKPOINT
          - 200 OK
          - 201 CREATED
          - 202 ACCEPTED
          - 203 NON_AUTHORITATIVE_INFORMATION
          - 204 NO_CONTENT
          - 205 RESET_CONTENT
          - 206 PARTIAL_CONTENT
          - 207 MULTI_STATUS
          - 208 ALREADY_REPORTED
          - 226 IM_USED
          - 300 MULTIPLE_CHOICES
          - 301 MOVED_PERMANENTLY
          - 302 FOUND
          - 302 MOVED_TEMPORARILY
          - 303 SEE_OTHER
          - 304 NOT_MODIFIED
          - 305 USE_PROXY
          - 307 TEMPORARY_REDIRECT
          - 308 PERMANENT_REDIRECT
          - 400 BAD_REQUEST
          - 401 UNAUTHORIZED
          - 402 PAYMENT_REQUIRED
          - 403 FORBIDDEN
          - 404 NOT_FOUND
          - 405 METHOD_NOT_ALLOWED
          - 406 NOT_ACCEPTABLE
          - 407 PROXY_AUTHENTICATION_REQUIRED
          - 408 REQUEST_TIMEOUT
          - 409 CONFLICT
          - 410 GONE
          - 411 LENGTH_REQUIRED
          - 412 PRECONDITION_FAILED
          - 413 PAYLOAD_TOO_LARGE
          - 413 REQUEST_ENTITY_TOO_LARGE
          - 414 URI_TOO_LONG
          - 414 REQUEST_URI_TOO_LONG
          - 415 UNSUPPORTED_MEDIA_TYPE
          - 416 REQUESTED_RANGE_NOT_SATISFIABLE
          - 417 EXPECTATION_FAILED
          - 418 I_AM_A_TEAPOT
          - 419 INSUFFICIENT_SPACE_ON_RESOURCE
          - 420 METHOD_FAILURE
          - 421 DESTINATION_LOCKED
          - 422 UNPROCESSABLE_ENTITY
          - 423 LOCKED
          - 424 FAILED_DEPENDENCY
          - 425 TOO_EARLY
          - 426 UPGRADE_REQUIRED
          - 428 PRECONDITION_REQUIRED
          - 429 TOO_MANY_REQUESTS
          - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
          - 451 UNAVAILABLE_FOR_LEGAL_REASONS
          - 500 INTERNAL_SERVER_ERROR
          - 501 NOT_IMPLEMENTED
          - 502 BAD_GATEWAY
          - 503 SERVICE_UNAVAILABLE
          - 504 GATEWAY_TIMEOUT
          - 505 HTTP_VERSION_NOT_SUPPORTED
          - 506 VARIANT_ALSO_NEGOTIATES
          - 507 INSUFFICIENT_STORAGE
          - 508 LOOP_DETECTED
          - 509 BANDWIDTH_LIMIT_EXCEEDED
          - 510 NOT_EXTENDED
          - 511 NETWORK_AUTHENTICATION_REQUIRED
      subErrors:
        type: array
        items:
          type: string
      timestamp:
        type: string
        example: dd-MM-yyyy hh:mm:ss
    title: ResponseError
  SummaryTable:
    type: object
    properties:
      Amount:
        type: string
      NumItems:
        type: string
    title: SummaryTable
  TextFormat:
    type: object
    properties:
      bold:
        type: string
      size:
        type: string
      underline:
        type: string
    title: TextFormat
  TotalTable:
    type: object
    properties:
      FeesAmount:
        type: string
      FeesPercentage:
        type: string
      InsuranceAmount:
        type: string
      InsuranceDamageAmount:
        type: string
      InsuranceEmploymentAmount:
        type: string
      InsuranceLifeAmount:
        type: string
    title: TotalTable

  AssetsCode:
    type: object
    properties:
      Mileage:
        type: string
      EstimatedAnnualMileage:
        type: string
      RegisterDate:
        type: string
      LicensePlate:
        type: string
      Vin:
        type: stringg
    title: AssetsCode

  Validation:
    type: object
    properties:
      MaxAmountFinanced:
        type: string
      MaxDeposit:
        type: string
      MaxPeriod:
        type: string
      MaxVehicleAgeAtEnd:
        type: string
      MaxVehicleAgeAtStart:
        type: string
      MinAmountFinanced:
        type: string
      MinDeposit:
        type: string
      MinPeriod:
        type: string
      MinVehicleAgeAtStart:
        type: string
      Period:
        type: string
      Rate:
        type: string
    title: Validation
    
parameters:
  AcceptParam:
    name: Accept
    type: string
    in: header
    description: 'The server response'
    required: true
    default: application/json
  AuthorizationParam:
    name: Authorization
    type: string
    in: header
    description: 'Authorization token. Bearer OAuth2 token'
    required: true
  ContentTypeParam:
    name: Content-Type
    type: string
    in: header
    description: 'metadata content type'
    required: true
    default: application/json
  ContextParam:
    name: Context
    type: string
    in: header
    description: 'include context object'
    required: true
  ParentSpanIdParam:
    name: X-B3-ParentSpanId
    type: string
    in: header
    description: 'Position of the parent operation in the trace tree. The value is 64 bits long. value is omitted when the span is the root of the trace tree. '
    required: false
  SampledParam:
    name: X-B3-Sampled
    type: string
    in: header
    description: 'Sampling decision. Sampling is a mechanism to reduce the volume of data in the tracing system. In B3, sampling applies consistently per-trace: once the sampling decision is made, the same value must be consistently sent downstream. This means that either all or no spans share a trace ID. The possible values are 0 = Deny 1 = Accept d = Debug'
    required: false
  SpanIdParam:
    name: X-B3-SpanId
    type: string
    in: header
    description: Position of the current operation in the trace tree. The value is 64 bits long. Do not integererpret the value it may or may not be derived from the value of the TraceId.
    required: false
  TraceIdParam:
    name: X-B3-TraceId
    type: string
    in: header
    description: Overall ID of the trace, shared by every span in the trace. The value is 64 or 128 bits long.
    required: false
  SantanderClientIdParam:
    name: X-Santander-Client-Id
    in: header
    description: Client Id header
    required: true
    type: string
  PasswordParam:
    name: Password
    type: string
    in: query
    description: Password
    required: true
    allowEmptyValue: false
  UserParam:
    name: User
    type: string
    in: query
    description: User
    required: true
    allowEmptyValue: false
  CountryParam:
    name: country
    in: path
    description: ES, UK, CH
    required: true
    type: string